requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
google-genai==0.8.0
openai==1.3.7
redis==5.0.1
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
eventlet==0.33.3
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
